VITIS-11024 enable hw_context support for xrt::graph objects 3rd commit
