+++ src/runtime_src/core/edge/user/aie/graph.cpp
@@ -16,14 +16,11 @@
  * under the License.
  */
 
-#include "graph.h"
-#ifndef __AIESIM__
 #include "core/edge/user/shim.h"
 #include "core/common/message.h"
 #ifndef __HWEM__
 #include "core/edge/user/plugin/xdp/aie_trace.h"
 #include "core/edge/user/plugin/xdp/aie_profile.h"
-#endif
 #endif
 #include "core/common/error.h"
 
@@ -38,342 +35,7 @@ extern "C"
 #include <xaiengine.h>
 }
 
-#ifdef __AIESIM__
-zynqaie::Aie* getAieArray()
-{
-  static zynqaie::Aie s_aie(xrt_core::get_userpf_device(0));
-  return &s_aie;
-}
-#endif
-
-namespace zynqaie {
-
-graph_instance::
-graph_instance(std::shared_ptr<xrt_core::device> dev, std::string graph_name,
-              xrt::graph::access_mode am, const zynqaie::hwctx_object* hwctx, const xrt_core::uuid uuid)
-  : device{std::move(dev)}, name{std::move(graph_name)}, m_hwctxHandle{hwctx}
-{
-    auto xclbin_uuid = m_hwctxHandle ? m_hwctxHandle->get_xclbin_uuid() : uuid;
-#ifndef __AIESIM__
-    auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
-
-    if (!drv->isAieRegistered())
-      throw xrt_core::error(-EINVAL, "No AIE presented");
-    aieArray = drv->getAieArray();
-#else
-    aieArray = getAieArray();
-#endif
-
-#ifndef __AIESIM__
-    id = xrt_core::edge::aie::get_graph_id(device.get(), name, m_hwctxHandle);
-    if (id == xrt_core::edge::aie::NON_EXIST_ID)
-        throw xrt_core::error(-EINVAL, "Can not get id for Graph '" + name + "'");
-
-    int ret = drv->openGraphContext(xclbin_uuid.get(), id, am);
-    if (ret)
-        throw xrt_core::error(ret, "Can not open Graph context");
-#endif
-    access_mode = am;
-
-    /* Initialize graph tile metadata */
-    graph_config = xrt_core::edge::aie::get_graph(device.get(), name, m_hwctxHandle);
-
-    /* Initialize graph rtp metadata */
-    rtps = xrt_core::edge::aie::get_rtp(device.get(), graph_config.id, m_hwctxHandle);
-
-    pAIEConfigAPI = std::make_shared<adf::graph_api>(&graph_config);
-    pAIEConfigAPI->configure();
-
-    state = graph_state::reset;
-#ifndef __AIESIM__
-    drv->getAied()->register_graph(this);
-#endif
-}
-
-graph_instance::
-~graph_instance()
-{
-#ifndef __AIESIM__
-    auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
-    if (!drv || !drv->getAied()) {
-	std::stringstream warnMsg;
-	warnMsg << "There is no active device open. Unable to close Graph `" << name << "`";
-	xrt_core::message::send(xrt_core::message::severity_level::warning, "XRT", warnMsg.str());
-	return;
-    }
-    drv->closeGraphContext(id);
-    drv->getAied()->deregister_graph(this);
-#endif
-}
-
-std::string
-graph_instance::
-getname() const
-{
-  return name;
-}
-
-unsigned short
-graph_instance::
-getstatus() const
-{
-  return static_cast<unsigned short>(state);
-}
-
-void
-graph_instance::
-reset()
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not reset graph");
-
-    for (int i = 0; i < graph_config.coreColumns.size(); i++) {
-      XAie_LocType coreTile = XAie_TileLoc(graph_config.coreColumns[i], graph_config.coreRows[i] + adf::config_manager::s_num_reserved_rows + 1);
-      XAie_CoreDisable(aieArray->getDevInst(), coreTile);
-    }
-
-    state = graph_state::reset;
-}
-
-uint64_t
-graph_instance::
-get_timestamp()
-{
-    /* TODO just use the first tile to get the timestamp? */
-    XAie_LocType coreTile = XAie_TileLoc(graph_config.coreColumns[0], graph_config.coreRows[0] + adf::config_manager::s_num_reserved_rows + 1);
-
-    uint64_t timeStamp;
-    AieRC rc = XAie_ReadTimer(aieArray->getDevInst(), coreTile, XAIE_CORE_MOD, &timeStamp);
-    if (rc != XAIE_OK)
-        throw xrt_core::error(-EINVAL, "Fail to read timestamp for Graph '" + name);
-
-    return timeStamp;
-}
-
-void
-graph_instance::
-run()
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not run graph");
-
-    if (state != graph_state::stop && state != graph_state::reset)
-      throw xrt_core::error(-EINVAL, "Graph '" + name + "' is already running or has ended");
-
-    pAIEConfigAPI->run();
-
-    state = graph_state::running;
-}
-
-void
-graph_instance::
-run(int iterations)
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not run graph");
-
-    if (state != graph_state::stop && state != graph_state::reset)
-      throw xrt_core::error(-EINVAL, "Graph '" + name + "' is already running or has ended");
-
-    pAIEConfigAPI->run(iterations);
-
-    state = graph_state::running;
-}
-
-void
-graph_instance::
-wait_done(int timeout_ms)
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not wait on graph");
-
-    if (state == graph_state::stop)
-      return;
-
-    if (state != graph_state::running)
-      throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not running, cannot wait");
-
-    auto begin = std::chrono::high_resolution_clock::now();
-
-    /*
-     * We are using busy waiting here. Until every tile in the graph
-     * is done, we keep polling each tile.
-     */
-    while (1) {
-        uint8_t done;
-        for (int i = 0; i < graph_config.coreColumns.size(); i++){
-            /* Skip multi-rate core */
-            if (graph_config.triggered[i]) {
-                done = 1;
-                continue;
-            }
-
-            XAie_LocType coreTile = XAie_TileLoc(graph_config.coreColumns[i], graph_config.coreRows[i] + adf::config_manager::s_num_reserved_rows + 1);
-            XAie_CoreReadDoneBit(aieArray->getDevInst(), coreTile, &done);
-            if (!done)
-                break;
-        }
-
-        if (done) {
-            state = graph_state::stop;
-            for (int i = 0; i < graph_config.coreColumns.size(); i++){
-                if (graph_config.triggered[i])
-                    continue;
-
-                XAie_LocType coreTile = XAie_TileLoc(graph_config.coreColumns[i], graph_config.coreRows[i] + adf::config_manager::s_num_reserved_rows + 1);
-                XAie_CoreDisable(aieArray->getDevInst(), coreTile);
-            }
-            return;
-        }
-
-        auto current = std::chrono::high_resolution_clock::now();
-        auto dur = current - begin;
-        auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(dur).count();
-        if (timeout_ms >= 0 && timeout_ms < ms)
-          throw xrt_core::error(-ETIME, "Wait graph '" + name + "' timeout.");
-    }
-}
-
-void
-graph_instance::
-wait()
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not wait on graph");
-
-    if (state == graph_state::stop)
-        return;
-
-    if (state != graph_state::running)
-        throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not running, cannot wait");
-
-    pAIEConfigAPI->wait();
-
-    state = graph_state::stop;
-}
-
-void
-graph_instance::
-wait(uint64_t cycle)
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not wait on graph");
-
-    if (state == graph_state::suspend)
-        return;
-
-    if (state != graph_state::running)
-        throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not running, cannot wait");
-
-    pAIEConfigAPI->wait(cycle);
-
-    state = graph_state::suspend;
-}
-
-void
-graph_instance::
-suspend()
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not suspend graph");
-
-    if (state != graph_state::running)
-      throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not running, cannot suspend");
-
-    for (int i = 0; i < graph_config.coreColumns.size(); i++) {
-        XAie_LocType coreTile = XAie_TileLoc(graph_config.coreColumns[i], graph_config.coreRows[i] + adf::config_manager::s_num_reserved_rows + 1);
-        XAie_CoreDisable(aieArray->getDevInst(), coreTile);
-    }
-
-    state = graph_state::suspend;
-}
-
-void
-graph_instance::
-resume()
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not resume on graph");
-
-    if (state != graph_state::suspend)
-      throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not suspended (wait(cycle)), cannot resume");
-
-    pAIEConfigAPI->resume();
-
-    state = graph_state::running;
-}
-
-void
-graph_instance::
-end()
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not end graph");
-
-    if (state != graph_state::running && state != graph_state::stop)
-      throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not running or stop, cannot end");
-
-    pAIEConfigAPI->end();
-
-    state = graph_state::end;
-}
-
-void
-graph_instance::
-end(uint64_t cycle)
-{
-    if (access_mode == xrt::graph::access_mode::shared)
-        throw xrt_core::error(-EPERM, "Shared context can not end graph");
-
-    if (state != graph_state::running && state != graph_state::suspend)
-        throw xrt_core::error(-EINVAL, "Graph '" + name + "' is not running or suspended, cannot end(cycle_timeout)");
-
-    pAIEConfigAPI->end(cycle);
-
-    state = graph_state::end;
-}
-
-
-void
-graph_instance::
-update_rtp(const std::string& port, const char* buffer, size_t size)
-{
-    auto it = rtps.find(port);
-    if (it == rtps.end())
-      throw xrt_core::error(-EINVAL, "Can't update graph '" + name + "': RTP port '" + port + "' not found");
-    auto& rtp = it->second;
-
-    if (access_mode == xrt::graph::access_mode::shared && !rtp.isAsync)
-        throw xrt_core::error(-EPERM, "Shared context can not update sync RTP");
-
-    if (rtp.isPL)
-      throw xrt_core::error(-EINVAL, "Can't update graph '" + name + "': RTP port '" + port + "' is not AIE RTP");
-
-    pAIEConfigAPI->update(&rtp, (const void*)buffer, size);
-}
-
-void
-graph_instance::
-read_rtp(const std::string& port, char* buffer, size_t size)
-{
-    auto it = rtps.find(port);
-    if (it == rtps.end())
-      throw xrt_core::error(-EINVAL, "Can't read graph '" + name + "': RTP port '" + port + "' not found");
-    auto& rtp = it->second;
-
-    if (rtp.isPL)
-      throw xrt_core::error(-EINVAL, "Can't read graph '" + name + "': RTP port '" + port + "' is not AIE RTP");
-
-    pAIEConfigAPI->read(&rtp, (void*)buffer, size);
-}
-
-} // zynqaie
-
-using graph_instance = zynqaie::graph_instance;
-
-namespace api {
-
-using graph_instance = zynqaie::graph_instance;
+typedef xclDeviceHandle xrtDeviceHandle;
 
 static inline
 std::string value_or_empty(const char* s)
@@ -381,10 +43,11 @@ std::string value_or_empty(const char* s)
   return s == nullptr ? "" : s;
 }
 
+namespace api {
+
 void
 xclAIEOpenContext(xclDeviceHandle handle, xrt::aie::access_mode am)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
@@ -393,22 +56,17 @@ xclAIEOpenContext(xclDeviceHandle handle, xrt::aie::access_mode am)
     throw xrt_core::error(ret, "Fail to open AIE context");
 
   drv->setAIEAccessMode(am);
-#endif
 }
 
 void
 xclSyncBOAIE(xclDeviceHandle handle, xrt::bo& bo, const char *gmioName, enum xclBOSyncDirection dir, size_t size, size_t offset)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
   if (!drv->isAieRegistered())
     throw xrt_core::error(-EINVAL, "No AIE presented");
   auto aieArray = drv->getAieArray();
-#else
-  auto aieArray = getAieArray();
-#endif
 
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
@@ -425,16 +83,12 @@ xclSyncBOAIE(xclDeviceHandle handle, xrt::bo& bo, const char *gmioName, enum xcl
 void
 xclSyncBOAIENB(xclDeviceHandle handle, xrt::bo& bo, const char *gmioName, enum xclBOSyncDirection dir, size_t size, size_t offset)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
   if (!drv->isAieRegistered())
     throw xrt_core::error(-EINVAL, "No AIE presented");
   auto aieArray = drv->getAieArray();
-#else
-  auto aieArray = getAieArray();
-#endif
 
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
@@ -451,16 +105,12 @@ xclSyncBOAIENB(xclDeviceHandle handle, xrt::bo& bo, const char *gmioName, enum x
 void
 xclGMIOWait(xclDeviceHandle handle, const char *gmioName)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
   if (!drv->isAieRegistered())
     throw xrt_core::error(-EINVAL, "No AIE presented");
   auto aieArray = drv->getAieArray();
-#else
-  auto aieArray = getAieArray();
-#endif
 
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
@@ -472,7 +122,6 @@ xclGMIOWait(xclDeviceHandle handle, const char *gmioName)
 void
 xclResetAieArray(xclDeviceHandle handle)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
@@ -483,17 +132,12 @@ xclResetAieArray(xclDeviceHandle handle)
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
   }
-
   aieArray->reset(device.get());
-#else
-  auto aieArray = getAieArray();
-#endif
 }
 
 int
 xclStartProfiling(xclDeviceHandle handle, int option, const char* port1Name, const char* port2Name, uint32_t value)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
@@ -501,9 +145,6 @@ xclStartProfiling(xclDeviceHandle handle, int option, const char* port1Name, con
     throw xrt_core::error(-EINVAL, "No AIE presented");
 
   auto aieArray = drv->getAieArray();
-#else
-  auto aieArray = getAieArray();
-#endif
 
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
@@ -515,7 +156,6 @@ xclStartProfiling(xclDeviceHandle handle, int option, const char* port1Name, con
 uint64_t
 xclReadProfiling(xclDeviceHandle handle, int phdl)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
@@ -523,9 +163,6 @@ xclReadProfiling(xclDeviceHandle handle, int phdl)
     throw xrt_core::error(-EINVAL, "No AIE presented");
 
   auto aieArray = drv->getAieArray();
-#else
-  auto aieArray = getAieArray();
-#endif
 
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
@@ -537,7 +174,6 @@ xclReadProfiling(xclDeviceHandle handle, int phdl)
 void
 xclStopProfiling(xclDeviceHandle handle, int phdl)
 {
-#ifndef __AIESIM__
   auto device = xrt_core::get_userpf_device(handle);
   auto drv = ZYNQ::shim::handleCheck(device->get_device_handle());
 
@@ -545,9 +181,6 @@ xclStopProfiling(xclDeviceHandle handle, int phdl)
     throw xrt_core::error(-EINVAL, "No AIE presented");
 
   auto aieArray = drv->getAieArray();
-#else
-  auto aieArray = getAieArray();
-#endif
 
   if (!aieArray->is_context_set()) {
     aieArray->open_context(device.get(), xrt::aie::access_mode::primary);
