+++ src/runtime_src/core/edge/user/aie/graph_object.h
@@ -1,14 +1,16 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2024 Advanced Micro Devices, Inc. All rights reserved.
+
 #ifndef _ZYNQ_GRAPH_OBJECT_H_
 #define _ZYNQ_GRAPH_OBJECT_H_
 
+#include "core/edge/user/aie/aie.h"
+#include "core/edge/user/aie/common_layer/adf_api_config.h"
 #include "core/common/message.h"
 #include "core/common/shim/shared_handle.h"
 #include "core/common/shim/graph_handle.h"
 #include "core/include/xrt/xrt_uuid.h"
 #include "xrt/xrt_graph.h"
-
 
 #include <memory>
 
@@ -17,16 +19,51 @@ namespace ZYNQ {
 }
 namespace zynqaie {
   // Shim handle for graph object
-  class graph_instance;
   class hwctx_object;
   class graph_object : public xrt_core::graph_handle
   {
     ZYNQ::shim* m_shim;
-    std::unique_ptr<zynqaie::graph_instance> m_graphInstance;
+    enum class graph_state : unsigned short
+    {
+      stop = 0,
+      reset = 1,
+      running = 2,
+      suspend = 3,
+      end = 4,
+    };
+    int id;
+    graph_state state;
+    std::string name;
+    xrt::graph::access_mode access_mode;
+
+    /**
+     * This is the pointer to the AIE array where the AIE part of
+     * the graph resides. The Aie is an obect that holds the whole
+     * AIE resources, configurations etc.
+     */
+    Aie* aieArray;
+
+    /**
+     * This is the collections of tiles that this graph uses.
+     * A tile is represented by a pair of number <col, row>
+     * It represents the tile position in the AIE array.
+     */
+    adf::graph_config graph_config;
+    std::shared_ptr<adf::graph_api> aie_config_api;
+    /* This is the collections of rtps that are used. */
+    std::unordered_map<std::string, adf::rtp_config> rtps;
 
   public:
     graph_object(ZYNQ::shim* shim, const xrt::uuid& uuid , const char* name,
                     xrt::graph::access_mode am, const zynqaie::hwctx_object* hwctx = nullptr);
+					
+	  ~graph_object();
+
+    std::string
+    getname() const;
+
+    unsigned short
+    getstatus() const;
 
     void
     reset_graph() override;
@@ -59,4 +96,4 @@ namespace zynqaie {
     read_graph_rtp(const char* port, char* buffer, size_t size) override;
   }; // graph_object
 }
-#endif
+#endif  //_ZYNQ_GRAPH_OBJECT_H_
