+++ src/runtime_src/core/pcie/emulation/sw_emu/generic_pcie_hal2/device_swemu.cxx
@@ -124,4 +124,78 @@ open_graph_handle(const xrt::uuid& xclbin_id, const char* name, xrt::graph::acce
                   static_cast<xclswemuhal2::SwEmuShim*>(get_device_handle()), xclbin_id, name, am);
 }
 
+#ifdef XRT_ENABLE_AIE
+ 
+void
+device::
+open_aie_context(xrt::aie::access_mode am)
+{
+ if (auto ret = xclAIEOpenContext(get_device_handle(), am))
+   throw error(ret, "fail to open aie context");
+}
+
+void
+device::
+sync_aie_bo(xrt::bo& bo, const char *gmioName, xclBOSyncDirection dir, size_t size, size_t offset)
+{
+  if (auto ret = xclSyncBOAIE(get_device_handle(), bo, gmioName, dir, size, offset))
+    throw system_error(ret, "fail to sync aie bo");
+}
+
+void
+device::
+reset_aie()
+{
+  if (auto ret = xclResetAIEArray(get_device_handle()))
+    throw system_error(ret, "fail to reset aie");
+}
+
+void
+device::
+sync_aie_bo_nb(xrt::bo& bo, const char *gmioName, xclBOSyncDirection dir, size_t size, size_t offset)
+{
+  if (auto ret = xclSyncBOAIENB(get_device_handle(), bo, gmioName, dir, size, offset))
+    throw system_error(ret, "fail to sync aie non-blocking bo");
+}
+
+void
+device::
+wait_gmio(const char *gmioName)
+{
+  if (auto ret = xclGMIOWait(get_device_handle(), gmioName))
+    throw system_error(ret, "fail to wait gmio");
+}
+
+int
+device::
+start_profiling(int option, const char* port1Name, const char* port2Name, uint32_t value)
+{
+  return xclStartProfiling(get_device_handle(), option, port1Name, port2Name, value);
+}
+
+uint64_t
+device::
+read_profiling(int phdl)
+{
+  return xclReadProfiling(get_device_handle(), phdl);
+}
+
+void
+device::
+stop_profiling(int phdl)
+{
+  if (auto ret = xclStopProfiling(get_device_handle(), phdl))
+    throw system_error(ret, "failed to stop profiling");
+}
+
+void
+device::
+load_axlf_meta(const axlf* buffer)
+{
+  if (auto ret = xclLoadXclBinMeta(get_device_handle(), buffer))
+    throw system_error(ret, "failed to load xclbin");
+}
+
+#endif
+
 }} // swemu,xrt_core
